test: outcomeMessage
