+++ src/components/questions.tsx
@@ -22,13 +22,7 @@ import {
   QuestionDetails as IQuestionDetails,
   RpcRemote,
 } from '../govTypes.js';
-
-enum Outcome {
-  // "win" just means some position won, including denial of a proposed change
-  Win = 'win',
-  // the election itself failed, for example because there was no quorum
-  Fail = 'fail',
-}
+import { Outcome, outcomeMessage } from '../lib/governance';
 
 export function OfferId(props: { id: number }) {
   const { id } = props;
@@ -40,31 +34,6 @@ export function OfferId(props: { id: number }) {
     console.debug('not a timestamp', id, err);
   }
   return <code title={title}>{id}</code>;
-}
-
-function outcomeMessage(outcome?: OutcomeRecord) {
-  if (!outcome) {
-    return '⏳ Vote Closes';
-  }
-
-  if (outcome.outcome === Outcome.Fail) {
-    return `❌ ${outcome.reason}`;
-  }
-
-  if (outcome.outcome === Outcome.Win) {
-    const positionKeys = Object.keys(outcome.position);
-    assert(
-      positionKeys.length === 1,
-      'Only single position outcomes supported'
-    );
-    const positionKey = positionKeys[0];
-    if (['noChange', 'dontUpdate'].includes(positionKey)) {
-      return `❌ Change Rejected`;
-    }
-    return `✅ Change Accepted`;
-  }
-
-  return `???`;
 }
 
 function outcomeColor(outcome?: OutcomeRecord) {
