--- src/test/stub.test.ts
@@ -1,5 +1,0 @@
-import { expect, it } from 'vitest';
-
-it('adds', () => {
-  expect(2 + 2).equals(4);
-});
